[Refactor] (code) Move the base classes 'BaseNode' and 'BasePath' to inner module 'model._data'.
