+++ smoothcrawler_cluster/_utils/zookeeper.py
@@ -13,140 +13,10 @@ from kazoo.client import KazooClient
 from kazoo.exceptions import NodeExistsError
 from kazoo.recipe.lock import ReadLock, Semaphore, WriteLock
 
+from ..model._data import BaseNode, MetaDataPath
 from .converter import BaseConverter
 
 BaseConverterType = TypeVar("BaseConverterType", bound=BaseConverter)
-
-
-class BaseNode(metaclass=ABCMeta):
-    """*Framework module to define some attributes of node in Zookeeper*
-
-    A node of Zookeeper.
-    """
-
-    @property
-    @abstractmethod
-    def path(self) -> str:
-        """:obj:`str`: Properties with both a getter and setter for the path of node in Zookeeper."""
-        pass
-
-    @path.setter
-    @abstractmethod
-    def path(self, val: str) -> None:
-        pass
-
-    @property
-    @abstractmethod
-    def value(self) -> str:
-        """:obj:`str`: Properties with both a getter and setter for the value of the path. It may need to deserialize
-        the data if it needs.
-        """
-        pass
-
-    @value.setter
-    @abstractmethod
-    def value(self, val: str) -> None:
-        pass
-
-
-class BasePath(metaclass=ABCMeta):
-    """*Base class to define all meta-data paths*
-
-    All the path properties of meta-data objects.
-    """
-
-    def __init__(self, name: str, group: str):
-        """
-
-        Args:
-            name (str): The name of current crawler instance.
-            group (str): The group what current crawler instance is in.
-        """
-        self._name = name
-        self._group = group
-
-    @property
-    def state_node_str(self) -> str:
-        """:obj:`str`: Property with only getter. The node naming of meta-data **GroupState** and **NodeState** path."""
-        return "state"
-
-    @property
-    @abstractmethod
-    def group_state(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **GroupState**."""
-        pass
-
-    @property
-    @abstractmethod
-    def node_state(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **NodeState**."""
-        pass
-
-    @property
-    def task_node_str(self) -> str:
-        """:obj:`str`: Property with only getter. The node naming of meta-data **Task** path."""
-        return "task"
-
-    @property
-    @abstractmethod
-    def task(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **Task**."""
-        pass
-
-    @property
-    def heartbeat_node_str(self) -> str:
-        """:obj:`str`: Property with only getter. The node naming of meta-data **Heartbeat** path."""
-        return "heartbeat"
-
-    @property
-    @abstractmethod
-    def heartbeat(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **Heartbeat**."""
-        pass
-
-
-class MetaDataPath(BasePath):
-    """*Base class to define all meta-data paths*
-
-    All the path properties of meta-data objects.
-    """
-
-    @property
-    def group_state(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **GroupState**."""
-        return f"{self.generate_parent_node(self._group, is_group=True)}/{self.state_node_str}"
-
-    @property
-    def node_state(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **NodeState**."""
-        return f"{self.generate_parent_node(self._name)}/{self.state_node_str}"
-
-    @property
-    def task(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **Task**."""
-        return f"{self.generate_parent_node(self._name)}/{self.task_node_str}"
-
-    @property
-    def heartbeat(self) -> str:
-        """:obj:`str`: Property with only getter. The path of meta-data **Heartbeat**."""
-        return f"{self.generate_parent_node(self._name)}/{self.heartbeat_node_str}"
-
-    @classmethod
-    def generate_parent_node(cls, parent_name: str, is_group: bool = False) -> str:
-        """Generate node path of Zookeeper with fixed format.
-
-        Args:
-            parent_name (str): The crawler name.
-            is_group (bool): If it's True, generate node path for _group_ type meta-data.
-
-        Returns:
-            str: A Zookeeper node path.
-
-        """
-        if is_group:
-            return f"group/{parent_name}"
-        else:
-            return f"node/{parent_name}"
 
 
 class ZookeeperPath(MetaDataPath):
